requirements.txt
fastapi==0.111.1
injector==0.22.0
python-dotenv==1.0.1
uvicorn==0.30.1
